old-tests/submission/Ms2ger/parsing/comment.dat

#data
<html><!-- comment --><title>Comment before head</title>
#errors
#document
| <html>
|   <!-- comment -->
|   <head>
|     <title>
|       "Comment before head"
|   <body>
